+++ test/flutter_fillbars_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-
 import 'package:flutter_fillbars/flutter_fillbars.dart';
 
 void main() {