+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.2
+
+- Added "background color" property which is used to set the color of the empty part of the bar
+- Added some border properties border:
+  - borderColor: sets the border's color.
+  - paddingColor: sets the color of the space between the fill area and the border
+  - externalMargin: sets the margin to other widgets
+  - borderPadding: sets the amount of padding between the border and the fillArea
+  - borderWidth: sets the width of the border itself
+  - radius: sets the Fillbar's border radius (0 = rectangular).
+
 ## 0.0.1
 
-- Added a basic static fillbar
+- Added a basic static fillbar with a value, width, height and a custom Color.